Finish question before 75;

+++ LeetCode/66.cpp
@@ -11,7 +11,7 @@ class Solution {
 public:
     vector<int> plusOne(vector<int>& digits) {
         int carry = 1;
-        for (int i = digits.size() - 1; i >= 0; i--) {
+        for (int i = (int)digits.size() - 1; i >= 0; i--) {
             digits[i] += carry;
             carry = digits[i] > 9 ? 1 : 0;
             digits[i] %= 10;
